Reduce SIGTERM timeout
Reduces the SIGTERM timeout to 0.5 seconds (from 60). Empirically, it
has been shown that there are no ill effects to a shorter timeout, and
it allows Pyrex to exit soon since there are some processes that like to
linger around for a long period of time before exiting (e.g. pseudo)