Use pyrex default values unless overridden by user
Instead of writing out the default user pyrex config file with all
values, write out a copy that has most values commented out. This means
that unless the user uncomments the variables, they will fall back to
the defaults specified in the default config file.

This is done by expanding the special token '%' in the config file to
either '' or '#' depending on how it is being used.