ci/test.py: Be more verbose
Travis builds are timing out because there is no output from the test
for 10 minutes. Convert the tests to no longer buffer output from
sub processes internally and set the PYREX_DOCKER_BUILD_QUIET variable
to 0 so that the tests are producing output while running.

Users can manually pass the "-b" option to the test script, which will
cause it to buffer all stdout and only print it if there is an error if
they want the less verbose behavior.