Make sure timezone package is installed
Some recipes (e.g. meta-mono) happen to rely on having a
valid timezone configured. The 'tzdata' package isn't
explicitly itemized on the set of required packages [1],
but the timezone data package is implicitly installed on
any physical workstation copy of Ubuntu.

It's probably not realistic to try to wean application build
systems away from use of packages that are conventionally
always on Linux, whether or not these appear formally in Yocto's
requirements.

[1] https://www.yoctoproject.org/docs/2.2/mega-manual/mega-manual.html#detailed-supported-distros