ci: Use BuildKit
Configured Travis to run a newer version of Docker that supports
BuildKit. BuildKit has many advantages:
 1) It is much faster to build in general
 2) It can build multi-stage images in parallel until dependencies are
    encountered
 3) It only builds the parts of a multistage build that are required to
    build the target image

All of these allow Travis to complete the CI jobs in a reasonable amount
of time (whereas a traditional image build would take far too long and
time out). It's so fast in fact that the build is no longer in danger of
exceeding the 10 minute no output timeout, so output buffering can be
re-enabled when running the test script, which significantly improves
the readability.