Add local rebuild detection
Adds support to try and determine if the locally built docker image
needs to be rebuilt when a pyrex command is run. Ideally, there would be
a way to ask docker about this, but currently this doesn't exist.
Instead, hash all of the input files for the docker image and verify the
hash when running a command.