docker: Add terminfo database
Adds the terminfo database to the docker images for maximal
compatibility with the host TERM setting. Adds a check and warning
message in entry.py if the user's $TERM value is unrecognized by the
container, and adds tests to ensure it all works as expected and that
there is a minimum set of TERM values supported.