Cleanup image name splitting
The image names currently take on the form NAME-VERSION-TYPE. There were
several places in the code where this was being split apart to grab one
or more fields. Improve these places to use a more pythonic and safer:

 (image_name, image_verison, image_type) = image.split('-')

This has the advantage of being simpler and also failing if the image
doesn't contain the required 3 fields.