ci: Unify image tests
Unifies the CI test script so that it will test all images at once. This
eliminates the need for setting the TEST_IMAGE environment variable to
select a test image and also makes it much easier for developers to test
all images (since that is now the default when the script is run).

This allows different image types to easily have a different set of
tests, eliminating the need to skip tests based on type.

Additionally, the developer documentation is updated with instructions
on how to run tests.