Rework capture behavior
Completely reworks the mechanism by which pyrex "capture" the build
environment. Now, instead of initializing in host environment, the
entire initialization and capture is done inside the target container,
in a container specific manner. This means that the host no longer needs
to have any of the dependencies and setup required for initialization,
since all of that is provided by the container.

Notable changes:
* The only valid mechanism for a user to provide a pyrex configuration
  is by setting the PYREXCONFFILE environment variable. The previous
  behaviors of copying template files no longer works.