Dockerfile: Build findmnt for Ubuntu 14.04
Builds newer util-linux for Ubuntu 14.04. This is primarily for the
purpose of providing a new enough findmnt command that correctly
supports the '--target' option required by Pyrex. util-linux doesn't
allow building only a static version of findmnt, so the entire
util-linux suite is built including setpriv, which requires libcap-ng to
also be built.