entry.py: Fix home directory safety check
Fixes the check to see if the home directory is safe to modify so that
is uses findmnt instead of checking the device node. The device node
will be the same for bind mounts originating from the same device so
it's not a safe check. Instead, instruct findmnt to find the mounted
device for the home directory. If it resolves to "/" it means that the
home directory isn't bound into the container and can be modified.