image: guard against userdel/groupdel failing on 0 input arguments
Depending on the details of what auxiliary groups or user accounts get
made during package installation, it can happen that there are no
non-root accounts or no non-root groups.

This trips up the logic that destroys all of those to present a clean
slate to the user account creation logic done by the entrypoint.

Teach xargs how to avoid the command entirely if there's no work to
be done.

Signed-off-by: Matt Hoosier <matt.hoosier@garmin.com>