workflows: run shellcheck for host and container scripts
All scripts except those in the `image` directory are supposed to be run
on the host. They are possibly run with zsh (or other shells) and not
only bash. Since bash supports `==` and not zsh nor in POSIX, let's try
to have the biggest compatibility by abiding by POSIX rules (except for
arrays, which are too convenient to disable).

SC1090[1] requires to provide a path to shellcheck so that it can source
the correct file and follow it. Since the path of sourced files are
possibly non-fixed, there is a need to disable this check.

SC2181[2] recommends to put the function directly in the if condition
rather than use $? but this makes the code less readable, so it makes
sense to disable it.

SC3054[3] pertains to array support in shell. Since bash and zsh are
supported, both of them supporting arrays, it is safe to disable it.

[1] https://github.com/koalaman/shellcheck/wiki/SC1090
[2] https://github.com/koalaman/shellcheck/wiki/SC2181
[3] https://github.com/koalaman/shellcheck/wiki/SC3054

Signed-off-by: Quentin Schulz <foss@0leil.net>