pyrex: define registry where the remote image is stored
When using podman and only passing a tag to `podman pull`, podman asks
the user which registry it should query for a given tag.

This makes it not very nice for automated scripts or user friendly
either.

Since garminpyrex images are stored on docker.io registry, let's default
to this value.

Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Signed-off-by: Quentin Schulz <foss@0leil.net>