Fix Let's Encrypt Root CA expiration
Works around the OpenSSL < 1.1.0 bug in Ubuntu Trusty that prevents it
from correctly handling the expired Root CA from Let's Encrypt.
Fortunately, all that needs to be done is to remove the expired
certificate and it will correctly detect the newer non-expired one.

Add a validate test to ensure this works correctly for all images.