ci/test.py: Fix race with bitbake exit outside of pyrex
When run outside of Pyrex, bitbake exits asynchronously which means that
it may still be cleaning up after the front-end `bitbake` command
returns. This can race with the directory deletion in test_top_dir, so
add a loop to wait up to 20 seconds for bitbake to be done before doing
the cleanup.