Allow run::bind config to specify different destination
Until now, the directories mentioned in this config key have always been
mounted at the same absolute path in the container as they exist on the
host. Extend this to allow a different destination path in the container:

  ...
  [run]
  bind =
    ...
    /src-path:/dst-path
    ...

All of the usual options already recognized are valid for entries
specified using this expanded syntax.

Change-Id: Iaa0e70325f576d71ea672da082ce1a76c9c81c94